# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API quản lý OpenVPN, Authentication, Portal...",
        "title": "System Portal API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/portal/audit/logs": {
            "get": {
                "parameters": [
                    {"type": "string", "in": "query", "name": "username"},
                    {"type": "string", "in": "query", "name": "group"},
                    {"type": "string", "in": "query", "name": "ip"},
                    {"type": "string", "in": "query", "name": "from"},
                    {"type": "string", "in": "query", "name": "to"},
                    {"type": "integer", "in": "query", "name": "page"},
                    {"type": "integer", "in": "query", "name": "limit"}
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit"
                ],
                "summary": "List audit logs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.AuditLog"
                            }
                        }
                    }
                }
            }
        },
        "/api/portal/audit/logs/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit"
                ],
                "summary": "Export audit logs",
                "responses": {
                    "200": {
                        "description": "not implemented",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/portal/audit/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit"
                ],
                "summary": "Audit statistics",
                "responses": {
                    "200": {
                        "description": "not implemented",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/portal/dashboard/activities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Recent activities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AuditResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/portal/dashboard/charts/activities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Activity chart data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/portal/dashboard/charts/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "User chart data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/portal/dashboard/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StatsResponse"
                        }
                    }
                }
            }
        },
        "/api/portal/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Groups"
                ],
                "summary": "List portal groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.PortalGroup"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Groups"
                ],
                "summary": "Create portal group",
                "parameters": [
                    {
                        "description": "Group data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.PortalGroup"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/portal/groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portal Groups"
                ],
                "summary": "Get portal group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.PortalGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/portal/permissions": {
            "get": {
                "security": [
                    {"BearerAuth": []}
                ],
                "produces": ["application/json"],
                "tags": ["Permissions"],
                "summary": "List permissions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {"$ref": "#/definitions/entities.Permission"}
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {"BearerAuth": []}
                ],
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Permissions"],
                "summary": "Create permission",
                "parameters": [
                    {
                        "description": "Permission data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {"$ref": "#/definitions/entities.Permission"}
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {"$ref": "#/definitions/response.SuccessResponse"}
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {"$ref": "#/definitions/response.ErrorResponse"}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuditResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.StatsResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "integer"
                }
            }
        },
...
        "entities.AuditLog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "userGroup": {
                    "type": "string"
                },
                "action": {
                    "type": "string"
                },
                "resourceType": {
                    "type": "string"
                },
                "resourceName": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                }
            }
        },
        "entities.Permission": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                }
            }
        },
        "entities.PortalGroup": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Permission"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "details": {},
                        "message": {
                            "type": "string"
                        },
                        "status": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "object",
                    "properties": {
                        "data": {},
                        "message": {
                            "type": "string"
                        },
                        "status": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}